rename: ChargerStateRedisService
